Move and rename mcg bucket replicaiton modules
