+++ ocs_ci/ocs/resources/mcg_bucket_replication/mockup_bucket_logger.py
